Refactor DSBSparse to DSDBSparse
- Updated all references from DSBSparse to DSDBSparse across multiple files to maintain consistency after renaming.
- Removed deprecated densify_selected_blocks function from sparse_utils.py.
- Ensured all documentation and type hints reflect the changes from DSBSparse to DSDBSparse.
